operator[] fix

+++ TicTacToeMinMax/TranspositionTable.h
@@ -19,7 +19,7 @@ public:
 
 	unsigned long long recalculateHash();
 	inline unsigned long long calculateHash(unsigned long long hash, int index) { return hash ^ zobristKeys[getKeyIndex(index)]; }
-	Entry& operator[](int h) { 
+	Entry& operator[](unsigned long long h) { 
 		if (entries[h % entries.size()].hash == h) return entries[h % entries.size()]; 
 		else return nullEntry; }
 	void placeEntry(unsigned long long hash, std::pair<int, int> scoreMove);
